+++ src/calculate/utils/create-result.ts
@@ -1,0 +1,16 @@
+import type { TimeCalculatorResult } from "../types";
+
+type TCreateResultArgs = {
+  result: string;
+  info?: string;
+};
+
+export function createResult({
+  result,
+  info,
+}: TCreateResultArgs): TimeCalculatorResult {
+  return {
+    result: result,
+    info: info,
+  };
+}