+++ src/calculate/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./format-date";
 export * from "./humanize-duration";
 export * from "./round-to-second";
+export * from "./create-result";